- Moved type declarations to global namespace - fixes #54
